requirements.txt
# python_version=3.9.7
kafka-python==2.0.2
findspark==2.0.1
pyspark==3.3.1
sdv==0.18.0
pyyaml